Feat: refactored paper and charsets
