+++ src/MSXCharacterSets.hh
@@ -1,0 +1,28 @@
+/** \file MSXCharacterSets.hh
+ * MSX Character Set Data for printers and plotters.
+ * Extracted from Printer.hh for shared use.
+ */
+
+#ifndef MSX_CHARACTER_SETS_HH
+#define MSX_CHARACTER_SETS_HH
+
+#include <cstdint>
+#include <span>
+
+namespace openmsx {
+
+// International MSX font (from NMS8250 BIOS ROM)
+[[nodiscard]] std::span<const uint8_t> getMSXFontRaw();
+
+// DIN MSX font (International with modified zero)
+[[nodiscard]] std::span<const uint8_t> getMSXDINFontRaw();
+
+// International MSX font, rotated 90 degrees and padded to 9 dots high.
+[[nodiscard]] std::span<const uint8_t> getMSXFontRotated();
+
+// Japanese MSX font (from Sony F1XV BIOS ROM)
+[[nodiscard]] std::span<const uint8_t> getMSXJPFontRaw();
+
+} // namespace openmsx
+
+#endif
