+++ src/meson.build
@@ -20,6 +20,7 @@ sources = files(
     'LedStatus.cc',
     'MSXBunsetsu.cc',
     'MSXCielTurbo.cc',
+    'MSXCharacterSets.cc',
     'MSXDevice.cc',
     'MSXDeviceSwitch.cc',
     'MSXE6Timer.cc',
@@ -38,6 +39,7 @@ sources = files(
     'MSXToshibaTcx200x.cc',
     'MSXTurboRPause.cc',
     'MSXVictorHC9xSystemControl.cc',
+    'Paper.cc',
     'PasswordCart.cc',
     'Pluggable.cc',
     'PluggableFactory.cc',
